pyproject.toml
[project]
name = "polybot"
version = "0.1.0"
description = ""
authors = [
    {name = "Astrid Klipfel",email = "astridklipfel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "emoji (>=2.14.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "polybot", from = "src"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
